WIP rename and move custom field options API hooks
