+++ front/app/api/custom_field_options/useDeleteCustomFieldOption.test.ts
@@ -4,27 +4,24 @@ import { setupServer } from 'msw/node';
 
 import createQueryClientWrapper from 'utils/testUtils/queryClientWrapper';
 
-import { userCustomFieldsOptionsData } from './__mocks__/useUserCustomFieldsOptions';
-import useAddUserCustomFieldOption from './useAddUserCustomFieldOption';
+import useDeleteCustomFieldOption from './useDeleteCustomFieldOption';
 
-const apiPath = '*/users/custom_fields/:customFieldId/custom_field_options';
+const apiPath =
+  '*/custom_field_options/:optionId';
 
 const server = setupServer(
-  http.post(apiPath, () => {
-    return HttpResponse.json(
-      { data: userCustomFieldsOptionsData },
-      { status: 200 }
-    );
+  http.delete(apiPath, () => {
+    return HttpResponse.json(null, { status: 200 });
   })
 );
 
-describe('useAddUserCustomFieldOption', () => {
+describe('useDeleteCustomFieldOption', () => {
   beforeAll(() => server.listen());
   afterAll(() => server.close());
 
   it('mutates data correctly', async () => {
     const { result, waitFor } = renderHook(
-      () => useAddUserCustomFieldOption(),
+      () => useDeleteCustomFieldOption(),
       {
         wrapper: createQueryClientWrapper(),
       }
@@ -32,24 +29,22 @@ describe('useAddUserCustomFieldOption', () => {
 
     act(() => {
       result.current.mutate({
-        customFieldId: 'customFieldId',
-        title_multiloc: { en: 'mock title' },
+        optionId: 'optionId'
       });
     });
 
     await waitFor(() => expect(result.current.isSuccess).toBe(true));
-    expect(result.current.data?.data).toEqual(userCustomFieldsOptionsData);
   });
 
   it('returns error correctly', async () => {
     server.use(
-      http.post(apiPath, () => {
+      http.delete(apiPath, () => {
         return HttpResponse.json(null, { status: 500 });
       })
     );
 
     const { result, waitFor } = renderHook(
-      () => useAddUserCustomFieldOption(),
+      () => useDeleteCustomFieldOption(),
       {
         wrapper: createQueryClientWrapper(),
       }
@@ -57,8 +52,7 @@ describe('useAddUserCustomFieldOption', () => {
 
     act(() => {
       result.current.mutate({
-        customFieldId: 'customFieldId',
-        title_multiloc: { en: 'mock title' },
+        optionId: 'optionId'
       });
     });
