+++ front/app/api/custom_field_options_weird/keys.ts
@@ -1,0 +1,13 @@
+import { QueryKeys } from 'utils/cl-react-query/types';
+
+const baseKey = { type: 'custom_field_option' };
+
+const customFieldOptionKeys = {
+  all: () => [baseKey],
+  items: () => [{ ...baseKey, operation: 'item' }],
+  item: (params: { id: string }) => [
+    { ...baseKey, operation: 'item', parameters: params },
+  ],
+} satisfies QueryKeys;
+
+export default customFieldOptionKeys;
