--- front/app/api/user_custom_fields_options/useReorderUserCustomFieldsOption.ts
@@ -1,36 +1,0 @@
-import { useMutation, useQueryClient } from '@tanstack/react-query';
-import { CLErrors } from 'typings';
-
-import fetcher from 'utils/cl-react-query/fetcher';
-
-import userCustomFieldsOptionsKeys from './keys';
-import { IUserCustomFieldOption } from './types';
-
-type IReorderCause = {
-  customFieldId: string;
-  optionId: string;
-  ordering: number;
-};
-
-const reorderOption = ({ customFieldId, optionId, ordering }: IReorderCause) =>
-  fetcher<IUserCustomFieldOption>({
-    path: `/users/custom_fields/${customFieldId}/custom_field_options/${optionId}/reorder`,
-    action: 'patch',
-    body: { custom_field_option: { ordering } },
-  });
-
-const useReorderUserCustomFieldOption = () => {
-  const queryClient = useQueryClient();
-  return useMutation<IUserCustomFieldOption, CLErrors, IReorderCause>({
-    mutationFn: reorderOption,
-    onSuccess: (_data, variables) => {
-      queryClient.invalidateQueries({
-        queryKey: userCustomFieldsOptionsKeys.list({
-          customFieldId: variables.customFieldId,
-        }),
-      });
-    },
-  });
-};
-
-export default useReorderUserCustomFieldOption;
