+++ front/app/containers/Admin/projects/project/analysis/Heatmap/HeatmapDetails/HeatmapTableHead.tsx
@@ -4,8 +4,8 @@ import { Thead, Tr, Th } from '@citizenlab/cl2-component-library';
 
 import { ICustomFieldBinData } from 'api/custom_field_bins/types';
 import useCustomFieldBins from 'api/custom_field_bins/useCustomFieldBins';
-import { IUserCustomFieldOptions } from 'api/user_custom_fields_options/types';
-import useCustomFieldsOptions from 'api/user_custom_fields_options/useUserCustomFieldsOptions';
+import { IUserCustomFieldOptions } from 'api/custom_field_options/types';
+import useCustomFieldsOptions from 'api/custom_field_options/useCustomFieldOptions';
 
 import { useGetOptionText } from './utils';
