+++ front/app/modules/commercial/representativeness/admin/utils/form.ts
@@ -1,6 +1,6 @@
 import { isEqual } from 'lodash-es';
 
-import { IUserCustomFieldOptionData } from 'api/user_custom_fields_options/types';
+import { IUserCustomFieldOptionData } from 'api/custom_field_options/types';
 
 import {
   IBinnedDistribution,
