+++ front/app/modules/commercial/smart_groups/components/UserFilterConditions/ValueSelector/CustomFieldOptionValuesSelector.tsx
@@ -3,8 +3,8 @@ import React from 'react';
 import { TRule } from 'modules/commercial/smart_groups/components/UserFilterConditions/rules';
 import { IOption } from 'typings';
 
-import { IUserCustomFieldOptionData } from 'api/user_custom_fields_options/types';
-import useUserCustomFieldOptions from 'api/user_custom_fields_options/useUserCustomFieldsOptions';
+import { IUserCustomFieldOptionData } from 'api/custom_field_options/types';
+import useUserCustomFieldOptions from 'api/custom_field_options/useCustomFieldOptions';
 
 import MultipleSelect from 'components/UI/MultipleSelect';
